requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.2
attrs==22.2.0
bidict==0.23.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
confection==0.0.4
configobj==5.0.8
configparser==6.0.0
cymem==2.0.7
distro==1.9.0
dnspython==2.7.0
et-xmlfile==1.1.0
etelemetry==0.3.0
eventlet==0.37.0
fastapi==0.115.4
ffmpeg-python==0.2.0
filelock==3.12.3
fitz==0.0.1.dev2
Flask==2.0.1
Flask-SocketIO==5.4.1
frontend==0.0.3
frozenlist==1.3.3
future==0.18.3
greenlet==3.1.1
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.0.1
joblib==1.3.2
langcodes==3.3.0
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.6.2
nose==1.3.7
numpy==1.25.2
openai==0.28.0
openpyxl==3.1.2
packaging==23.0
pandas==2.1.0
pandas-stubs==2.0.3.230814
pathlib==1.0.1
pathy==0.10.1
pillow==11.0.0
preshed==3.0.8
prov==2.0.0
pydantic==1.10.7
pydot==1.4.2
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing==3.1.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.17
python-socketio==5.11.4
pytz==2023.3
pyxnat==1.6
rdflib==7.0.0
regex==2023.3.22
requests==2.28.2
scipy==1.11.2
simple-websocket==1.1.0
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.36
srsly==2.4.6
starlette==0.41.2
thinc==8.1.9
tornado==6.2
tqdm==4.65.0
traits==6.3.2
typer==0.7.0
types-pytz==2023.3.0.1
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.23.2
wasabi==1.1.1
websockets==13.1
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
